test(e2e): extract HA clock-skew suite and split node-killing HA test
- Extract the four Clock Skew / Timezone Safety assertions out of the
  5-node docker HA suite (composed/ha/e2e_ha_full.parallel.test.ts) into a
  new in-process multi-node/high-availability/clock_skew.test.ts. They test
  PostgreSQL DB-clock semantics (timestamp storage + CURRENT_TIMESTAMP-based
  cleanup, immune to node clock skew), so they run against PGlite (real
  Postgres in WASM) rather than the mock-gossip stack, whose shared DB is
  driven by the node clock. Requires adding @electric-sql/pglite to end-to-end.
- Split the node-killing "distribute work" HA test into its own
  e2e_ha_distribute_work.parallel.test.ts, sharing the cluster setup via a new
  ha_full_setup.ts module. This removes the "must run last" ordering contract:
  it gets its own cluster, and the remaining e2e_ha_full tests are
  order-independent. No assertions changed in either file.
- Leave the two stranded composed tests in place, still excluded from CI:
  integration_proof_verification's committed epoch-proof fixture is stale (the
  proof no longer verifies) and e2e_persistence's beforeAll no longer completes
  (single-node sequencer stalls in checkpoint proposal). Both root causes are
  outside this diff; documented in the files and bootstrap.sh.
